backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-jose[cryptography]==3.3.0
sqlmodel==0.0.14
python-dotenv==1.0.1
pytest==7.0.0
pytest-asyncio==0.23.4

